+++ internal/server/handler.go
@@ -1,0 +1,19 @@
+package server
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+)
+
+func (s *Server) HelloWorldHandler(c echo.Context) error {
+	resp := map[string]string{
+		"message": "Hello World",
+	}
+
+	return c.JSON(http.StatusOK, resp)
+}
+
+func (s *Server) healthHandler(c echo.Context) error {
+	return c.JSON(http.StatusOK, s.db.Health())
+}
