+++ internal/server/routes.go
@@ -32,15 +32,3 @@ func (s *Server) RegisterRoutes() http.Handler {
 
 	return e
 }
-
-func (s *Server) HelloWorldHandler(c echo.Context) error {
-	resp := map[string]string{
-		"message": "Hello World",
-	}
-
-	return c.JSON(http.StatusOK, resp)
-}
-
-func (s *Server) healthHandler(c echo.Context) error {
-	return c.JSON(http.StatusOK, s.db.Health())
-}
